refactor: ObjectBox -> Sqlite로 전환 및 복약일정 작업 끝

+++ lib/presentation/bloc/page_b_circulator/page_b_circulator_cubit.dart
@@ -49,7 +49,7 @@ class PageBCirculatorCubit extends Cubit<PageBCirculatorState> {
     }
   }
 
-  void updateAlbumin(int? albumin) {
+  void updateAlbumin(double? albumin) {
     if (albumin == null) {
       emit(
         PageBCirculatorState(
